feat: a collapsible, reusable session inspector (#538)
The right sidebar is a session's whole "glance" surface and it keeps
growing: six stacked groups, most of them below the fold on any session
with a full metadata block. #486 hit exactly this and pinned the action
strip *outside* the scrolling pane; collapsing the data groups is the
other half of the same problem, and it is what makes a seventh group an
append rather than a layout decision.

Three parts.

`InspGroup` gains an **opt-in** `collapsible` / `open` / `onToggle`.
Eight views render it, so a default-collapsible group would silently
change every inspector in the app — `SaveBar`'s icon and error tone one
level down. A caller passing nothing takes an early return and emits the
markup it always did; the emitted-CSS diff (sorted set *and* ordered
selector list, the `charts.tsx` / `SaveBar.tsx` method) is exactly two
added `.insp-group__title--toggle*` rules, with every pre-existing rule
byte-identical and in place. The header is a real `button` with
`aria-expanded`, so Tab reaches it and Enter/Space toggle it with no
keydown handler here; `base.css` already resets a button's font, colour,
background, border and outline and puts it in the #469 selection
denylist, so the CSS is a box reset and deliberately carries no
`cursor: pointer` and no `:focus-visible` rule.

`lib/inspectorPrefs.ts` persists the toggles. It keys on a stable slug
rather than the rendered title — two of the groups interpolate a count
into their heading — and pins both the ids and the shipped defaults
through `lib/typeAssert.ts`, since a respelled id silently resets every
user's saved state and there is no test harness to catch it. Every key
is optional on read, so a blob from an older build keeps what it has.

`views/sessions/SessionInspector.tsx` is the extracted component, taking
`{ session }` and no handlers. `sessionMetrics.ts` beside it holds the
totals and the mode tables, because the table's Cost column and its row
badges call them too and exporting them from the view would have made
the view and the inspector import each other.

The detail page needs no second mount: `SessionDetail` fills
`.pane-detail` and the `inspectorOpen` `aside.pane-inspector` is its
sibling in the same `.panes`, so the Transcript and Journey tabs already
render this pane from this code.

Closes #538
